Move `ocr` and `match_text` into `_stbt/ocr.py`
Similar to fdca78f (Move `wait_for_motion` and `detect_motion` into
`motion.py`).
