+++ _stbt/core.py
@@ -12,13 +12,10 @@ from __future__ import absolute_import
 
 import argparse
 import datetime
-import errno
 import functools
 import inspect
 import itertools
 import os
-import re
-import subprocess
 import sys
 import threading
 import traceback
@@ -26,17 +23,14 @@ import warnings
 import weakref
 from collections import deque, namedtuple
 from contextlib import contextmanager
-from distutils.version import LooseVersion
 from textwrap import dedent
 
 import cv2
 import gi
 import numpy
-from enum import IntEnum
-from kitchen.text.converters import to_bytes
 
 import _stbt.cv2_compat as cv2_compat
-from _stbt import imgproc_cache, logging, utils
+from _stbt import imgproc_cache, logging
 from _stbt.config import ConfigurationError, get_config
 from _stbt.gst_utils import (array_from_sample, gst_iterate,
                              gst_sample_make_writable)
@@ -283,72 +277,6 @@ class _IsScreenBlackResult(object):
         return ("_IsScreenBlackResult(black=%r, frame=%s)" % (
             self.black,
             _frame_repr(self.frame)))
-
-
-class OcrMode(IntEnum):
-    """Options to control layout analysis and assume a certain form of image.
-
-    For a (brief) description of each option, see the `tesseract(1)
-    <https://github.com/tesseract-ocr/tesseract/blob/3.04.01/doc/tesseract.1.asc#options>`__
-    man page.
-    """
-    ORIENTATION_AND_SCRIPT_DETECTION_ONLY = 0
-    PAGE_SEGMENTATION_WITH_OSD = 1
-    PAGE_SEGMENTATION_WITHOUT_OSD_OR_OCR = 2
-    PAGE_SEGMENTATION_WITHOUT_OSD = 3
-    SINGLE_COLUMN_OF_TEXT_OF_VARIABLE_SIZES = 4
-    SINGLE_UNIFORM_BLOCK_OF_VERTICALLY_ALIGNED_TEXT = 5
-    SINGLE_UNIFORM_BLOCK_OF_TEXT = 6
-    SINGLE_LINE = 7
-    SINGLE_WORD = 8
-    SINGLE_WORD_IN_A_CIRCLE = 9
-    SINGLE_CHARACTER = 10
-
-    # For nicer formatting of `ocr` signature in generated API documentation:
-    def __repr__(self):
-        return str(self)
-
-
-class TextMatchResult(object):
-    """The result from `match_text`.
-
-    :ivar float time: The time at which the video-frame was captured, in
-        seconds since 1970-01-01T00:00Z. This timestamp can be compared with
-        system time (``time.time()``).
-
-    :ivar bool match: True if a match was found. This is the same as evaluating
-        ``MatchResult`` as a bool. That is, ``if result:`` will behave the same
-        as ``if result.match:``.
-
-    :ivar Region region: Bounding box where the text was found, or ``None`` if
-        the text wasn't found.
-
-    :ivar Frame frame: The video frame that was searched, as given to
-        `match_text`.
-
-    :ivar unicode text: The text that was searched for, as given to
-        `match_text`.
-    """
-    def __init__(self, time, match, region, frame, text):
-        self.time = time
-        self.match = match
-        self.region = region
-        self.frame = frame
-        self.text = text
-
-    # pylint: disable=E1101
-    def __nonzero__(self):
-        return self.match
-
-    def __repr__(self):
-        return (
-            "TextMatchResult(time=%s, match=%r, region=%r, frame=%s, "
-            "text=%r)" % (
-                "None" if self.time is None else "%.3f" % self.time,
-                self.match,
-                self.region,
-                _frame_repr(self.frame),
-                self.text))
 
 
 def new_device_under_test_from_config(
@@ -649,87 +577,6 @@ class DeviceUnderTest(object):
                     i += 1
                 else:
                     raise
-
-    def ocr(self, frame=None, region=Region.ALL,
-            mode=OcrMode.PAGE_SEGMENTATION_WITHOUT_OSD,
-            lang=None, tesseract_config=None, tesseract_user_words=None,
-            tesseract_user_patterns=None, upsample=True, text_color=None,
-            text_color_threshold=None):
-
-        if frame is None:
-            frame = self.get_frame()
-
-        if region is None:
-            raise TypeError(
-                "Passing region=None to ocr is deprecated since v0.21. "
-                "In a future version, region=None will mean an empty region "
-                "instead. To OCR an entire video frame, use "
-                "`region=Region.ALL`.")
-
-        if lang is None:
-            lang = get_config("ocr", "lang", "eng")
-
-        if isinstance(tesseract_user_words, (str, unicode)):
-            tesseract_user_words = [tesseract_user_words]
-
-        if isinstance(tesseract_user_patterns, (str, unicode)):
-            tesseract_user_patterns = [tesseract_user_patterns]
-
-        text, region = _tesseract(
-            frame, region, mode, lang, tesseract_config,
-            tesseract_user_patterns, tesseract_user_words, upsample, text_color,
-            text_color_threshold)
-        text = text.strip().translate(_ocr_transtab)
-        debug(u"OCR in region %s read '%s'." % (region, text))
-        return text
-
-    def match_text(self, text, frame=None, region=Region.ALL,
-                   mode=OcrMode.PAGE_SEGMENTATION_WITHOUT_OSD, lang=None,
-                   tesseract_config=None, case_sensitive=False,
-                   upsample=True, text_color=None, text_color_threshold=None):
-
-        import lxml.etree
-        if frame is None:
-            frame = self.get_frame()
-        if lang is None:
-            lang = get_config("ocr", "lang", "eng")
-
-        _config = dict(tesseract_config or {})
-        _config['tessedit_create_hocr'] = 1
-        if _tesseract_version() >= LooseVersion('3.04'):
-            _config['tessedit_create_txt'] = 0
-
-        rts = getattr(frame, "time", None)
-
-        xml, region = _tesseract(frame, region, mode, lang, _config,
-                                 None, text.split(), upsample, text_color,
-                                 text_color_threshold)
-        if xml == '':
-            result = TextMatchResult(rts, False, None, frame, text)
-        else:
-            hocr = lxml.etree.fromstring(xml.encode('utf-8'))
-            p = _hocr_find_phrase(hocr, text.split(), case_sensitive)
-            if p:
-                # Find bounding box
-                box = None
-                for _, elem in p:
-                    box = Region.bounding_box(box, _hocr_elem_region(elem))
-                # _tesseract crops to region and scales up by a factor of 3 so
-                # we must undo this transformation here.
-                n = 3 if upsample else 1
-                box = Region.from_extents(
-                    region.x + box.x // n, region.y + box.y // n,
-                    region.x + box.right // n, region.y + box.bottom // n)
-                result = TextMatchResult(rts, True, box, frame, text)
-            else:
-                result = TextMatchResult(rts, False, None, frame, text)
-
-        if result.match:
-            debug("match_text: Match found: %s" % str(result))
-        else:
-            debug("match_text: No match found: %s" % str(result))
-
-        return result
 
     def frames(self, timeout_secs=None):
         if timeout_secs is not None:
@@ -2211,266 +2058,3 @@ def _log_match_image_debug(imglog):
                 x._first_pass_matched for x in imglog.data["matches"]),  # pylint:disable=protected-access
             template_name=imglog.data["template_name"],
         ))
-
-
-# Tesseract sometimes has a hard job distinguishing certain glyphs such as
-# ligatures and different forms of the same punctuation.  We strip out this
-# superfluous information improving matching accuracy with minimal effect on
-# meaning.  This means that stbt.ocr give much more consistent results.
-_ocr_replacements = {
-    # Ligatures
-    u'ﬀ': u'ff',
-    u'ﬁ': u'fi',
-    u'ﬂ': u'fl',
-    u'ﬃ': u'ffi',
-    u'ﬄ': u'ffl',
-    u'ﬅ': u'ft',
-    u'ﬆ': u'st',
-    # Punctuation
-    u'“': u'"',
-    u'”': u'"',
-    u'‘': u'\'',
-    u'’': u'\'',
-    # These are actually different glyphs!:
-    u'‐': u'-',
-    u'‑': u'-',
-    u'‒': u'-',
-    u'–': u'-',
-    u'—': u'-',
-    u'―': u'-',
-}
-_ocr_transtab = dict((ord(amb), to) for amb, to in _ocr_replacements.items())
-
-
-def _find_tessdata_dir():
-    from distutils.spawn import find_executable
-
-    tessdata_prefix = os.environ.get("TESSDATA_PREFIX", None)
-    if tessdata_prefix:
-        tessdata = tessdata_prefix + '/tessdata'
-        if os.path.exists(tessdata):
-            return tessdata
-        else:
-            raise RuntimeError('Invalid TESSDATA_PREFIX: %s' % tessdata_prefix)
-
-    tess_prefix_share = os.path.normpath(
-        find_executable('tesseract') + '/../../share/')
-    for suffix in [
-            '/tessdata', '/tesseract-ocr/tessdata', '/tesseract/tessdata']:
-        if os.path.exists(tess_prefix_share + suffix):
-            return tess_prefix_share + suffix
-    raise RuntimeError('Installation error: Cannot locate tessdata directory')
-
-
-def _symlink_copy_dir(a, b):
-    """Behaves like `cp -rs` with GNU cp but is portable and doesn't require
-    execing another process.  Tesseract requires files in the "tessdata"
-    directory to be modified to set config options.  tessdata may be on a
-    read-only system directory so we use this to work around that limitation.
-    """
-    from os.path import basename, join, relpath
-    newroot = join(b, basename(a))
-    for dirpath, dirnames, filenames in os.walk(a):
-        for name in dirnames:
-            if name not in ['.', '..']:
-                rel = relpath(join(dirpath, name), a)
-                os.mkdir(join(newroot, rel))
-        for name in filenames:
-            rel = relpath(join(dirpath, name), a)
-            os.symlink(join(a, rel), join(newroot, rel))
-
-_memoise_tesseract_version = None
-
-
-def _tesseract_version(output=None):
-    r"""Different versions of tesseract have different bugs.  This function
-    allows us to tell the user if what they want isn't going to work.
-
-    >>> (_tesseract_version('tesseract 3.03\n leptonica-1.70\n') >
-    ...  _tesseract_version('tesseract 3.02\n'))
-    True
-    """
-    global _memoise_tesseract_version
-    if output is None:
-        if _memoise_tesseract_version is None:
-            try:
-                _memoise_tesseract_version = subprocess.check_output(
-                    ['tesseract', '--version'], stderr=subprocess.STDOUT)
-            except OSError as e:
-                if e.errno == errno.ENOENT:
-                    return None
-                else:
-                    raise
-        output = _memoise_tesseract_version
-
-    line = [x for x in output.split('\n') if x.startswith('tesseract')][0]
-    return LooseVersion(line.split()[1])
-
-
-def _tesseract(frame, region, mode, lang, _config, user_patterns, user_words,
-               upsample, text_color, text_color_threshold):
-
-    if _config is None:
-        _config = {}
-
-    if text_color_threshold is None:
-        text_color_threshold = get_config(
-            "ocr", "text_color_threshold", type_=int)
-
-    frame_region = _image_region(frame)
-    intersection = Region.intersect(frame_region, region)
-    if intersection is None:
-        warn("Requested OCR in region %s which doesn't overlap with "
-             "the frame %s" % (str(region), frame_region))
-        return (u'', None)
-    else:
-        region = intersection
-
-    return (_tesseract_subprocess(crop(frame, region), mode, lang, _config,
-                                  user_patterns, user_words, upsample,
-                                  text_color, text_color_threshold),
-            region)
-
-
-@imgproc_cache.memoize({"tesseract_version": str(_tesseract_version()),
-                        "version": "29"})
-def _tesseract_subprocess(
-        frame, mode, lang, _config, user_patterns, user_words, upsample,
-        text_color, text_color_threshold):
-
-    if upsample:
-        # We scale image up 3x before feeding it to tesseract as this
-        # significantly reduces the error rate by more than 6x in tests.  This
-        # uses bilinear interpolation which produces the best results.  See
-        # http://stb-tester.com/blog/2014/04/14/improving-ocr-accuracy.html
-        outsize = (frame.shape[1] * 3, frame.shape[0] * 3)
-        frame = cv2.resize(frame, outsize, interpolation=cv2.INTER_LINEAR)
-
-    if text_color is not None:
-        # Calculate distance of each pixel from `text_color`, then discard
-        # everything further than `text_color_threshold` distance away.
-        diff = numpy.subtract(frame, text_color, dtype=numpy.int32)
-        frame = numpy.sqrt((diff[:, :, 0] ** 2 +
-                            diff[:, :, 1] ** 2 +
-                            diff[:, :, 2] ** 2) / 3) \
-                     .astype(numpy.uint8)
-        _, frame = cv2.threshold(frame, text_color_threshold, 255,
-                                 cv2.THRESH_BINARY)
-
-    # $XDG_RUNTIME_DIR is likely to be on tmpfs:
-    tmpdir = os.environ.get("XDG_RUNTIME_DIR", None)
-
-    # The second argument to tesseract is "output base" which is a filename to
-    # which tesseract will append an extension. Unfortunately this filename
-    # isn't easy to predict in advance across different versions of tesseract.
-    # If you give it "hello" the output will be written to "hello.txt", but in
-    # hOCR mode it will be "hello.html" (tesseract 3.02) or "hello.hocr"
-    # (tesseract 3.03). We work around this with a temporary directory:
-    with utils.named_temporary_directory(prefix='stbt-ocr-', dir=tmpdir) as tmp:
-        outdir = tmp + '/output'
-        os.mkdir(outdir)
-
-        cmd = ["tesseract", '-l', lang, tmp + '/input.png',
-               outdir + '/output', "-psm", str(int(mode))]
-
-        tessenv = os.environ.copy()
-
-        if _config or user_words or user_patterns:
-            tessdata_dir = tmp + '/tessdata'
-            os.mkdir(tessdata_dir)
-            _symlink_copy_dir(_find_tessdata_dir(), tmp)
-            tessenv['TESSDATA_PREFIX'] = tmp + '/'
-
-        if user_words:
-            if 'user_words_suffix' in _config:
-                raise ValueError(
-                    "You cannot specify 'user_words' and " +
-                    "'_config[\"user_words_suffix\"]' at the same time")
-            with open('%s/%s.user-words' % (tessdata_dir, lang), 'w') as f:
-                f.write('\n'.join(to_bytes(x) for x in user_words))
-            _config['user_words_suffix'] = 'user-words'
-
-        if user_patterns:
-            if 'user_patterns_suffix' in _config:
-                raise ValueError(
-                    "You cannot specify 'user_patterns' and " +
-                    "'_config[\"user_patterns_suffix\"]' at the same time")
-            if _tesseract_version() < LooseVersion('3.03'):
-                raise RuntimeError(
-                    'tesseract version >=3.03 is required for user_patterns.  '
-                    'version %s is currently installed' % _tesseract_version())
-            with open('%s/%s.user-patterns' % (tessdata_dir, lang), 'w') as f:
-                f.write('\n'.join(to_bytes(x) for x in user_patterns))
-            _config['user_patterns_suffix'] = 'user-patterns'
-
-        if _config:
-            with open(tessdata_dir + '/configs/stbtester', 'w') as cfg:
-                for k, v in _config.iteritems():
-                    if isinstance(v, bool):
-                        cfg.write(('%s %s\n' % (k, 'T' if v else 'F')))
-                    else:
-                        cfg.write("%s %s\n" % (k, to_bytes(v)))
-            cmd += ['stbtester']
-
-        cv2.imwrite(tmp + '/input.png', frame)
-        subprocess.check_output(cmd, stderr=subprocess.STDOUT, env=tessenv)
-        with open(outdir + '/' + os.listdir(outdir)[0], 'r') as outfile:
-            return outfile.read().decode('utf-8')
-
-
-def _hocr_iterate(hocr):
-    started = False
-    need_space = False
-    for elem in hocr.iterdescendants():
-        if elem.tag == '{http://www.w3.org/1999/xhtml}p' and started:
-            yield (u'\n', elem)
-            need_space = False
-        if elem.tag == '{http://www.w3.org/1999/xhtml}span' and \
-                'ocr_line' in elem.get('class').split() and started:
-            yield (u'\n', elem)
-            need_space = False
-        for e, t in [(elem, elem.text), (elem.getparent(), elem.tail)]:
-            if t:
-                if t.strip():
-                    if need_space and started:
-                        yield (u' ', None)
-                    need_space = False
-                    yield (unicode(t).strip(), e)
-                    started = True
-                else:
-                    need_space = True
-
-
-def _hocr_find_phrase(hocr, phrase, case_sensitive):
-    if case_sensitive:
-        lower = lambda s: s
-    else:
-        lower = lambda s: s.lower()
-
-    words_only = [(lower(w).translate(_ocr_transtab), elem)
-                  for w, elem in _hocr_iterate(hocr) if w.strip() != u'']
-    phrase = [lower(_to_unicode(w)).translate(_ocr_transtab) for w in phrase]
-
-    # Dumb and poor algorithmic complexity but succint and simple
-    if len(phrase) <= len(words_only):
-        for x in range(0, len(words_only)):
-            sublist = words_only[x:x + len(phrase)]
-            if all(w[0] == p for w, p in zip(sublist, phrase)):
-                return sublist
-    return None
-
-
-def _to_unicode(text):
-    if isinstance(text, str):
-        return text.decode("utf-8")
-    else:
-        return unicode(text)
-
-
-def _hocr_elem_region(elem):
-    while elem is not None:
-        m = re.search(r'bbox (\d+) (\d+) (\d+) (\d+)', elem.get('title') or u'')
-        if m:
-            extents = [int(x) for x in m.groups()]
-            return Region.from_extents(*extents)
-        elem = elem.getparent()
