+++ tests/test-stbt-py.sh
@@ -316,8 +316,8 @@ test_press_visualisation() {
 }
 
 test_clock_visualisation() {
-    PYTHONPATH="$srcdir" python -c "import stbt, _stbt.core, distutils, sys; \
-        sys.exit(0 if (_stbt.core._tesseract_version() \
+    PYTHONPATH="$srcdir" python -c "import stbt, _stbt.ocr, distutils, sys; \
+        sys.exit(0 if (_stbt.ocr._tesseract_version() \
                        >= distutils.version.LooseVersion('3.03')) else 77)"
     case $? in
         0) true;;
